refactor: move mapping index helpers into utils
- keep workflow step definitions focused on orchestration

- expose cache parameter helpers from workflow_pathoscope.utils for tests
